requirements.txt
attrs==20.3.0
Automat==20.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
constantly==15.1.0
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
flask-mongoengine==1.0.0
Flask-WTF==0.14.3
gunicorn==20.0.4
heroku==0.1.4
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mongoengine==0.22.1
PyHamcrest==2.0.2
pymongo==3.11.2
python-dateutil==1.5
python-dotenv==0.15.0
requests==2.25.1
six==1.15.0
srv==1.1.0
Twisted==20.3.0
urllib3==1.26.3
Werkzeug==1.0.1
WTForms==2.3.3
zope.interface==5.2.0